feat: full project audit fixes, Playwright E2E, MD reorganization
- Fix mojibake (Dashboard, ActionHistory) — proper UTF-8 Polish characters
- Fix silent catches (AppContext, TrendExplorer, Alerts)
- Fix OAUTHLIB conditional in auth.py
- Centralize shared styles (TH_STYLE, MODAL) in UI.jsx
- Replace off-palette colors in Campaigns with v2 design system
- Delete dead code (Charts.jsx, SegmentBadge.jsx)
- Add Playwright E2E smoke tests (19 tests, mocked API)
- Add MarkdownComponents for Agent/Reports markdown rendering
- Archive 5 obsolete MD files to docs/archive/
- Update roadmap with implementation status (6 DONE, 4 PARTIAL, 16 NOT DONE)
- Sync COMPLETED_FEATURES.md with 6 newly delivered features
- Update SOURCE_OF_TRUTH.md, AGENTS.md references

+++ backend/app/schemas/ad.py
@@ -18,7 +18,7 @@ class AdResponse(BaseModel):
     impressions: int = 0
     cost_micros: int = 0
     conversions: float = 0
-    ctr: int = 0
+    ctr: float = 0.0
     created_at: Optional[datetime] = None
     updated_at: Optional[datetime] = None
